In platform agnostic code, use int_fast64_t instead of the posix type ssize_t

+++ src/realm/array_binary.cpp
@@ -60,7 +60,7 @@ void ArrayBinary::set(size_t ndx, BinaryData value, bool add_zero_term)
     size_t stored_size = value.size();
     if (add_zero_term)
         ++stored_size;
-    ssize_t diff =  (start + stored_size) - current_end;
+    int_fast64_t diff =  (start + stored_size) - current_end;
     m_blob.replace(start, current_end, value.data(), value.size(), add_zero_term);
     m_offsets.adjust(ndx, m_offsets.size(), diff);
